+++ dc-cudami-admin/src/main/resources/messages.properties
@@ -145,7 +145,9 @@ lbl.roles=Role(s)
 lbl.root_webpages=Root webpages
 lbl.row_number=#
 lbl.status=Status
+lbl.sub_webpages=Sub webpages
 lbl.tags=Tags
+lbl.text=Long text
 lbl.tooltip=Tooltip
 lbl.url=URL
 lbl.url_aliases=URL Aliases
@@ -210,6 +212,7 @@ page.title.predicates=Relation Types
 page.title.rendering_templates=Design templates
 page.title.topics=Topics
 page.title.users=Users
+page.title.webpages=Webpages
 page.title.websites=Websites
 page.title.works=Works
 parent_collections=Parent collections
