+++ dc-cudami-admin/src/main/resources/messages_de.properties
@@ -145,7 +145,9 @@ lbl.roles=Rolle(n)
 lbl.root_webpages=Webseiten auf oberster Ebene
 lbl.row_number=#
 lbl.status=Status
+lbl.sub_webpages=Untergeordnete Webseiten
 lbl.tags=Tags
+lbl.text=Langtext
 lbl.tooltip=Tooltip
 lbl.url=URL
 lbl.url_aliases=URL Aliasse
@@ -210,6 +212,7 @@ page.title.predicates=Beziehungstypen
 page.title.rendering_templates=Gestaltungsvorlagen
 page.title.topics=Themen
 page.title.users=Benutzer
+page.title.webpages=Webseiten
 page.title.websites=Websites
 page.title.works=Werke
 parent_collections=Übergeordnete Sammlungen
